requirements.txt
mysql-connector-python
python-telegram-bot
slack_sdk
pytz